pyharp/requirements.txt
gradio
descript-audiotools
symusic
-e git+https://github.com/TEAMuP-dev/pyharp.git#egg=pyharp
